+++ src/microservices/Microservices.FileCopier/Program.cs
@@ -16,7 +16,7 @@ namespace Microservices.FileCopier
             return Parser.Default.ParseArguments<CliOptions>(args).MapResult(
                 (o) =>
                 {
-                    GlobalOptions options = GlobalOptions.Load(o);
+                    GlobalOptions options = new GlobalOptionsFactory().Load(o);
 
                     var bootstrapper = new MicroserviceHostBootstrapper(() => new FileCopierHost(options));
                     return bootstrapper.Main();